pyproject.toml
[project]
name = "webtools"
version = "0.1.0"
description = "Tool that allows cloude models use the internet"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firecrawl-py>=1.13.2",
    "mcp[cli]>=1.3.0",
    "openai>=1.65.2",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
]
